refact/fix/feat: adicionadas validações, exibição de mensagens em cada estado

+++ scripts/tasks.js
@@ -1,3 +1,14 @@
+window.onload = function () {
+  listtask();
+};
+
+function timeMensagem(){
+  setTimeout(e => {
+    document.getElementsByClassName("mensagem")[0].innerHTML = ``
+    listtask();
+  }, 4000)
+}
+
 
 function listtask() {
   let url = `https://ctd-todo-api.herokuapp.com/v1/tasks`;
@@ -21,8 +32,9 @@ function listtask() {
           schemaA.innerHTML += `
                     <li class="tarefa">
                         <div class="not-done">
-                        <button type="button" onclick="editTaskStatus(${e.id}, true)"><i class="fas fa-check"></i>
-                        </button>
+                          <button type="button" onclick="editTaskStatus(${e.id}, true)">
+                          <span class="material-icons">schedule</span>
+                          </button>
                         </div>
 
                         <div class="descricao">
@@ -34,16 +46,19 @@ function listtask() {
         } else {
           schemaB.innerHTML += `
           <li class="tarefa">
-          <div class="done"></div>
-          <div class="descricao">
-          <p class="nome" id="desc_${e.id}">${e.description}</p>
-          <div>
-          <button onclick="editTaskStatus(${e.id},false)"><i id="${e.id}" class="fas
-          fa-undo-alt change"></i></button>
-          <button onclick="deleteTask(${e.id})"><i id="" class="far
-          fa-trash-alt"></i></button>
-          </div>
-          </div>
+            <div class="done">
+              <span class="material-icons">check</span>
+            </div>
+
+            <div class="descricao">
+              <p class="nome" id="desc_${e.id}">${e.description}</p>
+              <div>
+                <button onclick="editTaskStatus(${e.id},false)"><i id="${e.id}" class="fas
+                fa-undo-alt change"></i></button>
+                <button onclick="deleteTask(${e.id})"><i id="" class="far
+                fa-trash-alt"></i></button>
+              </div>
+            </div>
           </li>
                 `;
         }
@@ -52,9 +67,8 @@ function listtask() {
     .catch((err) => console.error(err));
 }
 
-window.onload = function () {
-  listtask();
-};
+
+
 
 function createTask() {
   let task = document.getElementById("novaTarea").value;
@@ -86,35 +100,44 @@ function createTask() {
     .catch((err) => console.error(err));
 }
 
-addEventListener("submit", (e) => {
-  e.preventDefault();
-});
-
 function editTaskStatus(id, status) {
-    let url = `https://ctd-todo-api.herokuapp.com/v1/tasks/${id}`;
-    let description = document.getElementById("desc_"+id).innerHTML;
-    let settings = {
-        method: "PUT",
-        headers: {
-          "Content-Type": "application/json",
-          Authorization: token,
-        },
-        body: JSON.stringify({
-          description: description,
-          completed: status,
-        }),
-      }
-      fetch(url, settings)
-    .then((response) => response.json())
-    .then((json) => {
-      console.log(json);
-      listtask();
-    })
-    .catch((err) => console.error(err));
+    if(status){
+      var txt = "Terminadas"
+    }else{
+      var txt = "Pendentes"
+    }
 
+    if(confirm(`Tem certeza que deseja mover a tafera para ${txt}?`)){
+      let url = `https://ctd-todo-api.herokuapp.com/v1/tasks/${id}`;
+      let description = document.getElementById("desc_"+id).innerHTML;
+      let settings = {
+          method: "PUT",
+          headers: {
+            "Content-Type": "application/json",
+            Authorization: token,
+          },
+          body: JSON.stringify({
+            description: description,
+            completed: status,
+          }),
+        }
+        fetch(url, settings)
+      .then((response) => response.json())
+      .then((json) => {
+        document.getElementsByClassName("mensagem")[0].innerHTML = `Tarefa <b>${json.description}</b> atualizada para <b>${txt}</b>!`
+        listtask();
+        timeMensagem();
+      })
+      .catch((err) => console.error(err));
+    }else{
+      document.getElementsByClassName("mensagem")[0].innerHTML = `A tarefa não foi atualizada!`
+      timeMensagem();
+    }
 }
 
 function deleteTask(id) {
+
+  if(confirm(`Tem certeza que deseja remover esta tarefa?`)){
     let url = `https://ctd-todo-api.herokuapp.com/v1/tasks/${id}`;
     let settings = {
         method: "DELETE",
@@ -127,8 +150,17 @@ function deleteTask(id) {
       fetch(url, settings)
     .then((response) => response.json())
     .then((json) => {
-      console.log(json);
+      document.getElementsByClassName("mensagem")[0].innerHTML = `A tarefa ${json.description} foi excluída!`
       listtask();
+      timeMensagem();
     })
     .catch((err) => console.error(err));
+  }else{
+    document.getElementsByClassName("mensagem")[0].innerHTML = `A tarefa não excluída!`
+    timeMensagem();
+  }
 }
+
+addEventListener("submit", (e) => {
+  e.preventDefault();
+});
